📝 chore(.gitignore): add env folder to .gitignore This commit adds the `env` folder to the `.gitignore` file in the `back-end` directory. The `env` folder contains environment variables and should not be committed to the repository.
🔥 refactor(HeroBanner.tsx): remove unused code
The code removed was unused and had no effect on the functionality of the component. The code removed included mouse event handlers and a distance function.

🎨 style(HeroBanner.tsx): add motion animation to hero banner text
The commit adds motion animation to the hero banner text to improve the user experience. The text now fades in and out and moves in different directions to create a more dynamic and engaging effect.

🎉 feat(scrollHoverMouse.tsx): add scroll and hover mouse effect
This commit adds a new component named scrollHoverMouse.tsx which provides a scroll and hover mouse effect. The component listens to mousemove and scroll events and sets the opacity and fade-in state of the component based on the distance of the mouse from the initial position and the scroll position.