🔍 chore(.gitignore): add .env file to gitignore 🎨 style(logo images): add new logo images to front-end assets The .env file is now added to the gitignore file, which prevents it from being committed to the repository. Two new logo images have been added to the front-end assets.
🎨 style(HeroBanner.tsx): update logo image
The logo image has been updated to a new version, which improves the visual appearance of the application. The old logo has been replaced with the new logo image.

🎨 style(HeroBanner.tsx): increase logo size and adjust text styles
The logo size has been increased to improve visibility and the text styles have been adjusted to improve readability and visual appeal. The first motion paragraph has been updated to reflect a new message, and the second motion paragraph has been updated to include a new span element. The third motion paragraph has been updated to include a new span element and adjust the text styles.

💄 style(HeroBanner.tsx): improve hero banner button style
The hero banner button has been updated to improve its style and user experience. The button is now centered and has a shadow effect when hovered over. The button text is now uppercase and the font size has been increased to improve readability.

🎨 style(index.css): refactor CSS animations and add new gradient background animation
This commit refactors the CSS animations by removing unused code and consolidating similar animations. A new gradient background animation has been added to the custom-animated-button class, which creates a smooth transition between three different opacity levels of the background color. The boxShadow class has been added to create a subtle shadow effect on elements.

✨ feat(fadeAnimations.css): add CSS animations for hero section
This commit adds CSS animations for the hero section of the website. The translateHeroInfos animation moves the hero section to the right and back to the left. The fadeIn, fadeInTwo, and fadeInThree animations fade in different elements of the hero section. The fadeOut animation fades out an element and moves it to the left. The animations are triggered on hover and have different durations and timing functions.